GDCDEVOPS-737 set up GDC Docker build
Add a Jenkinsfile for building the ProteinPaint backend Docker image.
Add an empty pre-commit file to support the build automation.

Add -b and -f flags to pack.sh so that the backend and frontend can be
built individually.

Pack the backend code inside the Docker build so that the Docker build
is entirely self-contained. To support this, remove the .git directory
from .dockerignore and add tmppack.

Make it possible to specify the source registry for the base Docker
image via the registry build arg. Default to Docker Hub as before.

Remove the public directory when building the Docker image so that when
ProteinPaint is run in backend-only mode, it won't have to remove that
directory on startup. Make it possible to run the container as an
unprivileged user that can't modify any files inside the container.

Note that as a side effect of some of the above changes, the Docker
image will only run in backend-only mode.